+++ requirements.txt
@@ -1,0 +1,3 @@
+tensorflow-gpu
+numpy
+matplotlib
