+++ src/controller.py
@@ -1,0 +1,11 @@
+class SimnetMNISTController:
+    pass
+
+class SimnetEMNISTController:
+    pass
+
+class DumbnetMNISTController:
+    pass
+
+class DumbnetEMNISTController:
+    pass
