+++ src/model.py
@@ -1,0 +1,7 @@
+from abc import ABC
+class Model(ABC):
+    pass
+
+class AbstractModel(Model):
+    pass
+
